Refactor: rename template to main.
